pyproject.toml
[project]
name = "dblp-semantic-processor"
version = "2.0.0"
description = "DBLP semantic data processing pipeline with scheduling and incremental updates"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28.0",
    "tqdm>=4.65.0", 
    "lxml>=4.9.0",
    "pandas>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "sqlalchemy>=2.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0", 
    "flake8>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88